Store events locally when db server is unavailable
After failing to store an event because of a database connection issue,
store events in a local h2 database in order to retain this data. The
default path to this database is in <gerrit_site>/events-db.

This same behaviour will also be present if the database cannot be
started upon plugin start or reload.

If the database server is disconnected and events are now being stored
locally, queries on the events shall return an HTTP error 503: Service
Unavailable.

Remove the configurable threadPoolSize and keep it hard-coded to 1. This
is not needed and will only cause problems if multiple threads try to
create the database simultaneously.

Change-Id: I1bb3fca2eeea4a37d74e32ae0f6910e33f2c84bb