Auto-recover from database connection loss
If the main database connection has been lost and events are being
stored locally, try to re-establish a connection in order to return to
normal mode of operation.

The connection check is scheduled in the WorkQueue so that the task may
be seen from show-queue. The time interval between which the plugin
should check the database is configurable in gerrit.config.

Once a connection has been established, return to normal operation and
restore events from the local database. Flush the back up database to
prevent duplication.

Change-Id: I3e43a66c2a2c8be4ed668033a82e832dec923c2c