Fail silently if there is no local db to restore
Upon starting the plugin, events are restored from a local database.
If the file did not exist, the error would be logged as a warning.
This warning should not be seen, since this is the main use case. The
local database file should only exist if there was a problem with the
main database and we are now recovering from that loss.

To reduce confusion, set the log levels to debug.

Change-Id: I72b5855302b89735c6a97eebe860c6d30cc4f3ad