When testing, be explicit for which projects to throw exception
While tests passed in Eclipse, they consistently failed when building
with Buck in tree on the command-line [1]. The output suggested that the
wrong event got killed. As the corresponding test relied on the order of
unordered events, we are now more explicit which event should cause
which reaction when setting up the mocks. Thereby, the tests now pass in
both Eclipse, and also for Buck.

[1] (wrapped for better readibility)
  FAILURE restoreFromLocalAndRemoveUnfoundProjectEvents: Not true that <[
    {"project":"mock project","type":"mock event","eventCreatedOn":1434569045},
    {"project":"unfound","type":"mock event","eventCreatedOn":1434569045}
  ]> contains exactly <[
    {"project":"mock project","type":"mock event","eventCreatedOn":1434569045},
    {"project":"proj","type":"mock event","eventCreatedOn":1434569045}
  ]>. It is missing <[
    {"project":"proj","type":"mock event","eventCreatedOn":1434569045}
  ]> and has unexpected items <[
    {"project":"unfound","type":"mock event","eventCreatedOn":1434569045}
  ]>

Change-Id: I46a897d59fc6ddc2d1ad638cf4e3d4bb2fdf01b3