Do not log exception stack trace when project not in database
If a project is present in the events-log database but it has been
deleted from Gerrit, a warning is logged stating this project will be
also deleted from the events-log database. Given the exception is
handled properly, i.e., a sensible action is taken, there is no need to
pollute the logs with a stack trace that does not provides any
additional useful information.

Change-Id: I5d30975440442b51bf99ea874f60d44f681d13ac