CheckConnectionTask: Fix log level of database connection check logs
When the plugin is used with default connection timeout (1000ms),
the "Checking database connection" message is emitted to the log
at level "info" every second. Reduce the level to debug.

The "Problem checking database connection" message is only emitted
at debug level, meaning in most installations it will not be seen.
Increase the level to error.

Change-Id: Iae48e56929b94857bdabdc0891440ebed1f69e72