Fix broken default configuration
The default url for the events-db was not correctly
resolving so the plugin could not find the database.

Also move the default store into the site.data_dir itself
rather than polluting the user's home folder. This also avoids
the case where the home folder may be not be able to be resolved.

Change-Id: I738c0fe37a39fce8388cd2b7c9b3f7c60aff38b7