SQLStoreTest: Reduce visibility of tearDown method
When running the tests by `bazel test //...` it fails with:

  [JUnit4TearDownNotRun] tearDown() method will not be run;
  please add JUnit's @After annotation. [1]

Adding the @After annotation as recommended is not the correct
solution, since this causes it to be run after each test which
results in failures due to the table being deleted.

Instead, change the method visibility to private so that bazel
doesn't expect it to be @After annotated.

[1] http://errorprone.info/bugpattern/JUnit4TearDownNotRun

Change-Id: I30450f7658db01f2f5ffb958b9edb06ad7b09c1c