Fix errorprone issue about tearDown method
Errorprone was giving the following error:tearDown() method will not be
run; please add JUnit's @After annotation.

In fact, the tearDown method was called manually in some tests so there
was no issue. But instead of renaming tearDown to something else, do the
same that was done on master in I2867bbd69 and convert tearDown to use
@After instead of expecting to be manually called.

Change-Id: I0a124946bb0c4d100e22e268c141d527285bff43