Periodically remove older database entries
So far, the removal of database entries older than the configured value
of 'maxAge' was done synchronously at the plugin startup. This approach
has two downsides: first, it can delay the plugin load and, second, if
the plugin is not reloaded in a long time (which is desirable), the old
entries accumulate on the database.

Implement a periodic task that removes old entries from the database
according to two parameters specified in the configuration: 'startTime'
and 'interval'. If any of these parameters is not present, old entries
are removed at plugin startup but asynchronously.

Change-Id: I19e3ca9149e93fdb02f8630182b8bd85843611f0