Replace Apache commons-dbcp with HikariCP
As some benchmarks show [1], [2], [3], HikariCP is a lot faster than the
old commons-dcbp version inherited from the core dependencies, and with
minimal configuration.

Besides that, the library avoids locks as much as possible and the code
base is small. Projects like Liferay [4], splunk, the Play framework and
even some Apache projects [5] use HikariCP.

[1] https://beansroasted.wordpress.com/tag/connection-pool-comparison/
[2] http://www.trustiv.co.uk/2014/06/battle-connection-pools
[3] https://github.com/brettwooldridge/HikariCP#jmh-benchmarks-checkered_flag
[4] https://community.liferay.com/blogs/-/blogs/tomcat-hikaricp
[5] https://github.com/brettwooldridge/HikariCP/blob/dev/documents/Wall-of-Fame.md

Change-Id: I79fe729e668fedf246f32b88a14a1c2c437cc8b0