Do not remove entries from local DB
The H2 based local DB serves as a backup storage when the main
events-log database is offline and it is emptied every time the former
is back online. This makes unnecessary to cleanup this local database.

Besides, if the local DB has never been necessary because the main one
has been online all the time, when trying to delete from the local one,
an error occurs complaining about non existent database.

Change-Id: Idf61bda5e9c5c791dae78166b5e8dd76ade8da19