Update patch set 2
Patch Set 2:

Okay, everything is formatted according to "google-java-format-1.6".

In addition, I restarted my Gerrit server a bunch and found out that MySQL's "CREATE INDEX" syntax is not idempotent, so I replaced it with a more complicated workflow that figures out if the index is present (storing the value in "@x") and builds a query (either a no-op "SELECT" or the "ALTER TABLE" query to add the index).

I tested on my server with a dozen or so restarts and also dropping the "created_idx" and "project_idx" indexes some of the times.  It creates the indexes just fine if they aren't there, and it doesn't encounter any errors if they already are.

Patch-set: 2
Reviewer: Gerrit User 1055293 <1055293@173816e5-2b9a-37c3-8a2e-48639d4f1153>