Platform-independent '--workspace_status_command' script
Rewrite tools/workspace-status.sh script on python to ensure
better compatibility with Windows.
(As suggested here: https://github.com/bazelbuild/bazel/issues/5958)

The solution requires a Python binary on the classpath, but
this should be the case most of the time.

The script should be both Python2 and Python3 compatible, since
"python" or "python.exe" binary usally link to Python2 on linux
and to Python3 on Windows.

Change-Id: I6935fa1a27803797b2470f770bc1b6abed2936b8