Update patch set 1
Patch Set 1:

(5 comments)

> Patch Set 1:
> 
> Should this be done on stable-2.14?
> 
> And should the same be done for the other plugins too?

I agree with David, that we should do it on stable-2.14 and do it in Gerrit core first. And when that change is accepted, then we could do it for otehr plugins, taht supprt stanalone build mode.

Note, that as a workaround you could just add workspace-status.py locally in addition to workspace-status.sh (and not instead of it) and run:

  $ bazel build --workspace_status_command="python tools/workspace-status.py" events-log

Patch-set: 1
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>