Make SQLEntry id an Object rather than int
Since the Spanner implementation uses String UUID for the primary
key, the SQLEntry id needs to accept both String and int values.

This change adapts the compareTo, equals, and hashCode methods as
well.

The previous hashCode method was not really hashing the value, so
I have updated that to use Objects.hashCode and removed the now-
unnecessary testHashCode() test.

The only change to the SQLClient is to use getObject instead of
getInt.

Change-Id: Id9e48ee897263012275d14ae7b5aa0952f11f43f