Insert event time as Instant for Spanner implementation
Google Cloud Spanner adds seven hours to java.sql.Timestamp objects
when converting to UTC, resulting in entries being timestamped as
occurring in the future.

I believe this is related to their treatment of Timestamp literals,
which assume a default time zone of America/Los_Angeles:
https://cloud.google.com/spanner/docs/reference/standard-sql/lexical#timestamp_literals

Using Instant as an explicit representation of UTC timestamps keeps
the Spanner entries accurate.

Change-Id: I313af4fbd7803478fb0604913ee2d781de37c895