Insert event time as Instant for Spanner implementation
Google Cloud Spanner adds seven hours to java.sql.Timestamp objects
when converting to UTC, resulting in entries being timestamped as
occurring in the future.

This results in entries being timestamped as occurring in the future.
Using Instant as an explicit representation of UTC timestamps keeps
the Spanner entries accurate.

Change-Id: I313af4fbd7803478fb0604913ee2d781de37c895